backend/pyproject.toml
[project]
name = "sniffnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "jupyterlab>=4.4.2",
    "kagglehub>=0.3.12",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "scikit-image>=0.25.2",
    "sqlalchemy>=2.0.41",
    "torchmetrics[image]>=1.7.1",
    "alembic>=1.16.2",
    "fastapi[standard]>=0.115.13",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.0",
    "python-jose>=3.5.0",
    "bcrypt==4.0.1",
    "uvicorn>=0.34.3",
    "pytest>=8.4.1",
    "kaggle>=1.8.2",
    "scikit-learn>=1.8.0",
    "requests>=2.32.0",
]

[project.scripts]
api = "sniffnet.api.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
